+++ apps/meteor/client/views/admin/rooms/RoomsTableFilters.tsx
@@ -2,7 +2,7 @@ import { Box, Icon, TextInput } from '@rocket.chat/fuselage';
 import type { OptionProp } from '@rocket.chat/ui-client';
 import { MultiSelectCustom } from '@rocket.chat/ui-client';
 import { useTranslation } from '@rocket.chat/ui-contexts';
-import React, { useCallback, useEffect, useState } from 'react';
+import React, { useCallback, useState } from 'react';
 import type { Dispatch, ReactElement, SetStateAction } from 'react';
 
 const roomTypeFilterStructure = [
@@ -12,12 +12,7 @@ const roomTypeFilterStructure = [
 		isGroupTitle: true,
 	},
 	{
-		id: 'channels',
-		text: 'Channels',
-		checked: false,
-	},
-	{
-		id: 'directMessages',
+		id: 'd',
 		text: 'Direct_Message',
 		checked: false,
 	},
@@ -27,8 +22,18 @@ const roomTypeFilterStructure = [
 		checked: false,
 	},
 	{
-		id: 'omnichannel',
+		id: 'l',
 		text: 'Omnichannel',
+		checked: false,
+	},
+	{
+		id: 'p',
+		text: 'Private_Channels',
+		checked: false,
+	},
+	{
+		id: 'c',
+		text: 'Public_Channels',
 		checked: false,
 	},
 	{
@@ -38,37 +43,28 @@ const roomTypeFilterStructure = [
 	},
 ] as OptionProp[];
 
-const roomVisibilityFilterStructure = [
-	{
-		id: 'filter_by_visibility',
-		text: 'Filter_by_visibility',
-		isGroupTitle: true,
-	},
-	{
-		id: 'private',
-		text: 'Private',
-		checked: false,
-	},
-	{
-		id: 'public',
-		text: 'Public',
-		checked: false,
-	},
-] as OptionProp[];
-
 const RoomsTableFilters = ({ setFilters }: { setFilters: Dispatch<SetStateAction<any>> }): ReactElement => {
 	const t = useTranslation();
 	const [text, setText] = useState('');
-	const [roomTypeOptions, setRoomTypeOptions] = useState<OptionProp[]>(roomTypeFilterStructure);
-	const [roomVisibilityOptions, setRoomVisibilityOptions] = useState<OptionProp[]>(roomVisibilityFilterStructure);
+
 	const [roomTypeSelectedOptions, setRoomTypeSelectedOptions] = useState<OptionProp[]>([]);
-	const [roomVisibilitySelectedOptions, setRoomVisibilitySelectedOptions] = useState<OptionProp[]>([]);
 
-	useEffect(() => {
-		return setFilters({ searchText: text, types: roomTypeSelectedOptions, visibility: roomVisibilitySelectedOptions });
-	}, [setFilters, roomTypeSelectedOptions, roomVisibilitySelectedOptions, text]);
+	const handleSearchTextChange = useCallback(
+		(event) => {
+			const text = event.currentTarget.value;
+			setFilters({ searchText: text, types: roomTypeSelectedOptions });
+			setText(text);
+		},
+		[roomTypeSelectedOptions, setFilters],
+	);
 
-	const handleSearchTextChange = useCallback((event) => setText(event.currentTarget.value), []);
+	const handleRoomTypeChange = useCallback(
+		(options: OptionProp[]) => {
+			setFilters({ searchText: text, types: options });
+			setRoomTypeSelectedOptions(options);
+		},
+		[text, setFilters],
+	) as Dispatch<SetStateAction<OptionProp[]>>;
 
 	return (
 		<Box
@@ -92,23 +88,11 @@ const RoomsTableFilters = ({ setFilters }: { setFilters: Dispatch<SetStateAction
 			</Box>
 			<Box minWidth='x224' m='x4'>
 				<MultiSelectCustom
-					dropdownOptions={roomTypeOptions}
+					dropdownOptions={roomTypeFilterStructure}
 					defaultTitle={'All_rooms' as any}
 					selectedOptionsTitle='Rooms'
-					setSelectedOptions={setRoomTypeSelectedOptions}
+					setSelectedOptions={handleRoomTypeChange}
 					selectedOptions={roomTypeSelectedOptions}
-					customSetSelected={setRoomTypeOptions}
-				/>
-			</Box>
-
-			<Box minWidth='x224' m='x4'>
-				<MultiSelectCustom
-					dropdownOptions={roomVisibilityOptions}
-					defaultTitle={'All_visible' as any}
-					selectedOptionsTitle='Visible'
-					setSelectedOptions={setRoomVisibilitySelectedOptions}
-					selectedOptions={roomVisibilitySelectedOptions}
-					customSetSelected={setRoomVisibilityOptions}
 				/>
 			</Box>
 		</Box>