+++ apps/meteor/client/views/admin/viewLogs/hooks/useAnalyticsObject.ts
@@ -1,0 +1,8 @@
+import { useEndpoint } from '@rocket.chat/ui-contexts';
+import { useQuery } from '@tanstack/react-query';
+
+export const useAnalyticsObject = () => {
+	const getAnalytics = useEndpoint('GET', '/v1/statistics');
+
+	return useQuery(['analytics'], () => getAnalytics({}), { staleTime: 10 * 60 * 1000 });
+};