+++ apps/meteor/server/models/raw/Statistics.ts
@@ -44,4 +44,25 @@ export class StatisticsRaw extends BaseRaw<IStats> implements IStatisticsModel {
 			},
 		);
 	}
+
+	async findMonthlyPeakConnections() {
+		const oneMonthAgo = new Date();
+		oneMonthAgo.setDate(oneMonthAgo.getDate() - 30);
+		oneMonthAgo.setHours(0, 0, 0, 0);
+
+		return this.findOne<Pick<IStats, 'dailyPeakConnections' | 'createdAt'>>(
+			{
+				createdAt: { $gte: oneMonthAgo },
+			},
+			{
+				sort: {
+					dailyPeakConnections: -1,
+				},
+				projection: {
+					dailyPeakConnections: 1,
+					createdAt: 1,
+				},
+			},
+		);
+	}
 }