+++ apps/meteor/server/services/omnichannel/service.ts
@@ -1,10 +1,21 @@
 import { ServiceClassInternal } from '@rocket.chat/core-services';
 import type { IOmnichannelService } from '@rocket.chat/core-services';
+import type { IOmnichannelQueue } from '@rocket.chat/core-typings';
 
-import { Livechat } from '../../../app/livechat/server';
+import { Livechat } from '../../../app/livechat/server/lib/LivechatTyped';
+import { RoutingManager } from '../../../app/livechat/server/lib/RoutingManager';
+import { settings } from '../../../app/settings/server';
+import { OmnichannelQueue } from './queue';
 
 export class OmnichannelService extends ServiceClassInternal implements IOmnichannelService {
 	protected name = 'omnichannel';
+
+	private queueWorker: IOmnichannelQueue;
+
+	constructor() {
+		super();
+		this.queueWorker = new OmnichannelQueue();
+	}
 
 	async created() {
 		this.onEvent('presence.status', async ({ user }): Promise<void> => {
@@ -17,5 +28,15 @@ export class OmnichannelService extends ServiceClassInternal implements IOmnicha
 				await Livechat.notifyAgentStatusChanged(user._id, user.status);
 			}
 		});
+	}
+
+	async started() {
+		settings.watch<boolean>('Livechat_enabled', (enabled) => {
+			void (enabled && RoutingManager.isMethodSet() ? this.queueWorker.shouldStart() : this.queueWorker.stop());
+		});
+	}
+
+	getQueueWorker(): IOmnichannelQueue {
+		return this.queueWorker;
 	}
 }