+++ apps/meteor/tests/end-to-end/api/livechat/19-business-hours.ts
@@ -22,12 +22,12 @@ import {
 	getDepartmentById,
 	deleteDepartment,
 } from '../../../data/livechat/department';
-import { createAgent, makeAgentAvailable } from '../../../data/livechat/rooms';
+import { createAgent, createManager, makeAgentAvailable } from '../../../data/livechat/rooms';
 import { removeAgent } from '../../../data/livechat/users';
 import { removePermissionFromAllRoles, restorePermissionToRoles, updateSetting, updateEESetting } from '../../../data/permissions.helper';
 import type { IUserCredentialsHeader } from '../../../data/user';
 import { password } from '../../../data/user';
-import { createUser, deleteUser, getMe, login } from '../../../data/users.helper';
+import { setUserActiveStatus, createUser, deleteUser, getMe, getUserByUsername, login } from '../../../data/users.helper';
 import { IS_EE } from '../../../e2e/config/constants';
 
 describe('LIVECHAT - business hours', function () {
@@ -249,7 +249,7 @@ describe('LIVECHAT - business hours', function () {
 		});
 	});
 
-	(IS_EE ? describe : describe.skip)('[EE] BH operations upon creation', () => {
+	(IS_EE ? describe : describe.skip)('[EE][BH] On Business Hour created', () => {
 		let defaultBusinessHour: ILivechatBusinessHour;
 
 		before(async () => {
@@ -288,7 +288,7 @@ describe('LIVECHAT - business hours', function () {
 	//  and "dep2" and both these depts are connected to same BH, then in this case after
 	//  archiving "dep1", we'd still need to BH within this user's cache since he's part of
 	//  "dep2" which is linked to BH
-	(IS_EE ? describe : describe.skip)('[EE] BH operations post department archiving', () => {
+	(IS_EE ? describe : describe.skip)('[EE][BH] On Department archived', () => {
 		let defaultBusinessHour: ILivechatBusinessHour;
 		let customBusinessHour: ILivechatBusinessHour;
 		let deptLinkedToCustomBH: ILivechatDepartment;
@@ -434,7 +434,7 @@ describe('LIVECHAT - business hours', function () {
 			await deleteUser(agentLinkedToDept.user);
 		});
 	});
-	(IS_EE ? describe : describe.skip)('[EE] BH operations post department disablement', () => {
+	(IS_EE ? describe : describe.skip)('[EE][BH] On Department disabled', () => {
 		let defaultBusinessHour: ILivechatBusinessHour;
 		let customBusinessHour: ILivechatBusinessHour;
 		let deptLinkedToCustomBH: ILivechatDepartment;
@@ -578,7 +578,7 @@ describe('LIVECHAT - business hours', function () {
 			await deleteUser(agentLinkedToDept.user);
 		});
 	});
-	(IS_EE ? describe : describe.skip)('[EE] BH operations post department removal', () => {
+	(IS_EE ? describe : describe.skip)('[EE][BH] On Department removed', () => {
 		let defaultBusinessHour: ILivechatBusinessHour;
 		let customBusinessHour: ILivechatBusinessHour;
 		let deptLinkedToCustomBH: ILivechatDepartment;
@@ -702,7 +702,7 @@ describe('LIVECHAT - business hours', function () {
 			await deleteUser(agentLinkedToDept.user);
 		});
 	});
-	describe('BH behavior upon new agent creation/deletion', () => {
+	describe('[CE][BH] On Agent created/removed', () => {
 		let defaultBH: ILivechatBusinessHour;
 		let agent: ILivechatAgent;
 		let agentCredentials: IUserCredentialsHeader;
@@ -780,6 +780,77 @@ describe('LIVECHAT - business hours', function () {
 
 		after(async () => {
 			await deleteUser(agent._id);
+		});
+	});
+
+	describe('[CE][BH] On Agent deactivated/activated', () => {
+		let defaultBH: ILivechatBusinessHour;
+		let agent: ILivechatAgent;
+
+		before(async () => {
+			await updateSetting('Livechat_enable_business_hours', true);
+			await updateEESetting('Livechat_business_hour_type', LivechatBusinessHourBehaviors.SINGLE);
+			// wait for callbacks to run
+			await sleep(2000);
+
+			defaultBH = await getDefaultBusinessHour();
+			await openOrCloseBusinessHour(defaultBH, true);
+
+			agent = await createUser();
+			await createAgent(agent.username);
+		});
+
+		after(async () => {
+			await deleteUser(agent);
+			await updateSetting('Livechat_enable_business_hours', false);
+		});
+
+		it('should verify if agent becomes unavailable to take chats when user is deactivated', async () => {
+			await setUserActiveStatus(agent._id, false);
+
+			const latestAgent = await getUserByUsername(agent.username);
+
+			expect(latestAgent).to.be.an('object');
+			expect(latestAgent.statusLivechat).to.be.equal(ILivechatAgentStatus.NOT_AVAILABLE);
+		});
+
+		it('should verify if agent becomes available to take chats when user is activated, if business hour is active', async () => {
+			await openOrCloseBusinessHour(defaultBH, true);
+
+			await setUserActiveStatus(agent._id, true);
+
+			const latestAgent = await getUserByUsername(agent.username);
+
+			expect(latestAgent).to.be.an('object');
+			expect(latestAgent.statusLivechat).to.be.equal(ILivechatAgentStatus.AVAILABLE);
+		});
+		it('should verify if agent becomes unavailable to take chats when user is activated, if business hour is inactive', async () => {
+			await openOrCloseBusinessHour(defaultBH, false);
+
+			await setUserActiveStatus(agent._id, false);
+			await setUserActiveStatus(agent._id, true);
+
+			const latestAgent = await getUserByUsername(agent.username);
+
+			expect(latestAgent).to.be.an('object');
+			expect(latestAgent.statusLivechat).to.be.equal(ILivechatAgentStatus.NOT_AVAILABLE);
+		});
+		it('should verify if managers are not able to make deactivated agents available', async () => {
+			await createManager();
+
+			await setUserActiveStatus(agent._id, false);
+
+			const response = await request
+				.post(api('livechat/agent.status'))
+				.set(credentials)
+				.send({
+					status: 'available',
+					agentId: agent._id,
+				})
+				.expect(400);
+
+			expect(response.body).to.have.property('success', false);
+			expect(response.body).to.have.property('error', 'error-user-deactivated');
 		});
 	});
 });