+++ ee/packages/ddp-client/src/types/streams.ts
@@ -1,6 +1,5 @@
 import type { AppStatus } from '@rocket.chat/apps-engine/definition/AppStatus';
 import type { ISetting as AppsSetting } from '@rocket.chat/apps-engine/definition/settings';
-import type { IUIKitInteraction } from '@rocket.chat/apps-engine/definition/uikit';
 import type {
 	IMessage,
 	IRoom,
@@ -24,6 +23,7 @@ import type {
 	ILivechatAgent,
 	IImportProgress,
 	IBanner,
+	UiKit,
 } from '@rocket.chat/core-typings';
 
 type ClientAction = 'inserted' | 'updated' | 'removed' | 'changed';
@@ -148,7 +148,7 @@ export interface StreamerEvents {
 		{ key: `${string}/notification`; args: [INotificationDesktop] },
 		{ key: `${string}/voip.events`; args: [VoipEventDataSignature] },
 		{ key: `${string}/call.hangup`; args: [{ roomId: string }] },
-		{ key: `${string}/uiInteraction`; args: [IUIKitInteraction] },
+		{ key: `${string}/uiInteraction`; args: [UiKit.ServerInteraction] },
 		{
 			key: `${string}/video-conference`;
 			args: [{ action: string; params: { callId: VideoConference['_id']; uid: IUser['_id']; rid: IRoom['_id'] } }];